Change in API, now we have a sinle input generator, not two different for bands and relax.
There is now the single BaseWorkchainInputsGenerator.
